% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nonwear.R
\name{check.nonwear}
\alias{check.nonwear}
\title{Check for Non-Wear in Accelerometer Data}
\usage{
check.nonwear(
  data,
  window = 3 * 94 * 60,
  tol = 10,
  tol.upper = 0,
  filter = F,
  minimum.wear.bout = 94 * 60 * 24
)
}
\arguments{
\item{data}{Data frame containing the raw accelerometer data. Should include variables time, x, y, z.}

\item{window}{Number of samples for minimum contiguous period of non-wear. To be passed to accelerometry::weartime.}

\item{tol}{Number of allowances in any window of non-wear. To be passed to accelerometry::weartime}

\item{tol.upper}{Maximum allowable sequential change in any window of non-wear. To be passed to accelerometry::weartime}

\item{filter}{Should nonwear be removed from the data set?}

\item{minimum.wear.bout}{Minimum contiguous period of wear time for a period to be considered wear.}
}
\value{
A data frame with variables time, x, y, z, wear, and wear.bout

- wear = an indicator for whether a person is wearing the device.

- wear.bout = an integer numbering which consecutive period of contiguous device wear a time point corresponds to. 0 indicates nonwear.
}
\description{
Checks for non-wear by checking for long intervals of no change in the signal along each axis.
}
