% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclid_norm.R
\name{euclid.norm}
\alias{euclid.norm}
\title{Euclidian Norm, Radius}
\usage{
euclid.norm(x, y, z, v = cbind(x, y, z))
}
\arguments{
\item{x}{First accelerometer channel.}

\item{y}{Second accelerometer channel.}

\item{z}{Third accelerometer channel.}

\item{v}{optional matrix or data frame containing x,y,z measurments.}
}
\description{
Calculate length of a 3D vector.
}
