% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_top.R
\name{find.top}
\alias{find.top}
\title{Estimate upright position from accelerometry data}
\usage{
find.top(data, p = 0.95)
}
\arguments{
\item{data}{a data frame with variables x,y,z}
}
\description{
Takes a data frame with variables x,y,z and returns a vector of length 3 indicating orientation of upright position.
}
