% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_activity_summaries.R
\name{global.activity.summaries}
\alias{global.activity.summaries}
\title{Calculate Common Activity Summaries from Summarised Data}
\usage{
global.activity.summaries(data, epoch.seconds = 60)
}
\arguments{
\item{data}{A data frame with epoch-level activity summaries. This function is designed to input the data frame returned from Rtbeat::process.zacl directly. If not, variables should include time (class dttm), mad (Vector magnitude counts), wear (indicator for whether the device was actually worn during a given period), and time.group (a count variable indicating the distinct wear bout).}

\item{epoch.seconds}{How many seconds does each row in the "data" argument correspond to?}
}
\value{
data frame with the following variables

- start: time of starting epoch

- end: time of last epoch

- epoch.seconds: duration of epoch-level summaries (same as input)

- days: duration of recording time in days

- wear.days: duration of wear time in days (should be less than days)

- wear.bout.count: number of wear bouts in the case of intermittant non-wear

- mad_0...mad_22: average mean absolute deviation of acceleration in two hour bins throughout the day.

- mad: daily average mean absolute deviation

- down: daily proportion of time spent recumbent
}
\description{
Takes a data frame with activity measurements (MAD), and time. Returns a data frame with one row with relevant global activity summaries.
}
