% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_alldays.R
\name{is_alldays}
\alias{is_alldays}
\title{Inter-Daily Stability}
\usage{
is_alldays(activity, time, bin_epochs, epoch_seconds)
}
\arguments{
\item{activity}{vector representing timeseries across all days of activity data.
Each entry corresponds to one epoch.}

\item{time}{same length as activity vector.}

\item{bin_epochs}{number of epochs per bin to collapse activity into before performing
IS formula.}
}
\description{
Calculate Inter-Daily Stability (IS) from one day of data using
a given number of epochs (usually minutes) per bin.
}
