% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_oneday.R
\name{iv_oneday}
\alias{iv_oneday}
\title{Intra-Daily Variability}
\usage{
iv_oneday(activity_oneday, bin_epochs)
}
\arguments{
\item{activity_oneday}{vector representing timeseries for one day of activity data.
Each entry corresponds to one epoch.}

\item{bin_epochs}{number of epochs per bin to collapse activity_oneday into before performing
IV formula.}
}
\description{
Calculate Intra-Daily Variability (IV) from one day of data using
a given number of epochs (usually minutes) per bin.
}
