% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad.R
\name{mad}
\alias{mad}
\title{Vector Magnitude Counts (Mean Absolute Deviation)}
\usage{
mad(x, y, z, roll = FALSE, window = 94)
}
\arguments{
\item{x}{First accelerometer channel.}

\item{y}{Second accelerometer channel.}

\item{z}{Third accelerometer channel.}

\item{roll}{Should mad be calculated using a sliding window?}

\item{window}{What number of observations should be used in the sliding window?
Default is 94, as Zio measures approximately 94 observations per minute.}
}
\description{
Calculate mean absolute deviation from triaxial accelerometer data for a collection of time points.
}
