% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_data.R
\name{rotate.data}
\alias{rotate.data}
\title{Rotate Triaxial Accelerometer Data}
\usage{
rotate.data(data, from, to)
}
\arguments{
\item{data}{Data frame or matrix containing the accelerometer data. Should include ONLY variables x, y, z.}

\item{from}{Vector defining the starting point for the rotation.}

\item{to}{Vector defining the ending point for the rotation.}
}
\description{
Takes a data frame or data matrix with three columns, and internally generates a rotation matrix that will rotate the vector "from" to "to". The vector space orthoganol to these vectors (their cross-product) are not rotated--they become the axis of rotation.
}
