% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_accel.R
\docType{data}
\name{sample.accel}
\alias{sample.accel}
\title{Sample Zio Accelerometer Data}
\format{
A  data frame with four columns including
\describe{
\item{time}{the associated date-time of the accelerometer measurement. The dates and timezones have been changed from their original versions to remove any participant identifying information.}
\item{x, y, z}{accelerometer measurements along three orthoganol axes in gravitational units. }
}
}
\source{
Data in this sample were collected by the Multicenter AIDS Cohort Study (MACS).
Dates were changed.
The MACS website is located at http://aidscohortstudy.org/.
}
\usage{
sample.accel
}
\description{
A sample accelerometer data file for one individual over the course of wear (~ 14 days). This data is used to illustrate processing of activity and posture summaries.
}
\references{
Amir S. Heravi, Lacey Etzkorn, Jacek Urbanek, Ciprian Crainiceanu, Naresh M. Pun- jabi, Hiroshi Ashikaga, Todd T. Brown, Matthew J. Budoff, Gypsyamber D’Souza, Jared W. Magnani, Frank J. Palella Jr., Ronald D. Berger, Katherine C. Wu, and Wendy S. Post. Hiv infection is associated with variability in ventricular repolariza- tion: The multicenter aids cohort study (macs). Circulation, 11 Nov 2019.
}
\keyword{datasets}
