% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_minutes.R
\docType{data}
\name{sample.minutes}
\alias{sample.minutes}
\title{Sample Zio Minute-Level Summaries}
\format{
A  data frame with minute level summaries. See `processl.zacl`.
}
\source{
Data in this sample were collected by the Multicenter AIDS Cohort Study (MACS).
All dates were changed.
The MACS website is located at http://aidscohortstudy.org/.
}
\usage{
data(grav)
}
\description{
A sample minute-level summary data file for one individual over the course of wear (~ 14 days).
This data was generated with `process.zacl(sample.accel, cluster = "none")`.
}
\keyword{datasets}
