% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_zacl.R
\name{process.zacl}
\alias{process.zacl}
\title{Process Zio ZACL File}
\usage{
process.zacl(
  data,
  epoch.seconds = 60,
  p = 0.95,
  k = 0.98,
  theta.star = 45,
  screen.nonwear = T,
  check.device.rotation = T,
  nonwear.window = 3 * 94 * 60,
  nonwear.tol = 10,
  nonwear.tol.upper = 0,
  minimum.wear.bout = 94 * 60 * 24,
  cluster = "none"
)
}
\arguments{
\item{data}{Data frame containing the accelerometer data. Should include variables time, x, y, z.}

\item{epoch.seconds}{Duration (seconds) of interval in which to summarise the data.
Options include 10, 30, 60, 300, or "none".
If epoch.seconds is "none", then data is processed at the original frequency of the data.#'}

\item{p}{percentile for defining high-acceleration point. Used in estimation of upright orientation.}

\item{k}{ratio cutoff for detecting removal/rotation/reattachment.}

\item{theta.star}{Threshold for inclination for determining whether a point is recumbent or upright.}

\item{check.device.rotation}{Boolean. Should device removal/rotation/reattachment be screened for?}

\item{nonwear.window}{Window length (units of sampling points on raw data scale) representing minimum possible
nonwear period. Passed to accelerometry::weartime. Defaults to 16920, about 3 hours of Zio wear time.}

\item{nonwear.tol}{Number of allowed changes in window length to be considered nonwear. Defaults to 10.}

\item{nonwear.tol.upper}{Maximum allowable sum of change in axis gravitational units in any window of non-wear.
To be passed to accelerometry::weartime.
Default of NULL is equivalent to zero.}

\item{minimum.wear.bout}{Window length (units of sampling points on raw data scale) representing minimum possible
wear period. Defaults to 135360, about 24 hours of Zio wear time.}

\item{cluster}{Clustering algorithm to create posture groups. Either "meanShift", "ward", "centroid", or "none".
Choosing "none" will dramatically reduce processing time.}

\item{check.nonwear}{Boolean. Should non-wear be screened for?}
}
\value{
A data frame with epoch level activity summaries.
For each time stamp in the resulting data frame, the summary corresponds to the following epoch.seconds
(i.e. 22/9/19 10:50:00 corresponds to the time interval from 22/9/19 10:50:00 to 22/9/19 10:54:59
if epoch.seconds=300).

\item{cluster}{Which posture cluster does an epoch belong to?}

\item{down}{Indicator for recumbent.}

\item{wear.bout}{Integer index for period of consecutive wear epochs an epoch corresponds to.}

\item{theta}{estimated inclination of the chest during the epoch.}

\item{mad}{Mean absolute deviation, a measure of activity intensity during the epoch (milli-gravitational units).}

\item{x,y,z}{Median accelerations measured along each axis during an epoch (gravitational units).}
}
\description{
Takes the raw Zio accelerometer file and processes it into adjacent intervals of length epoch.seconds.
}
